feat: rename to string_op
