+++ datafusion/functions-aggregate-common/src/aggregate/groups_accumulator/string_op.rs
